+++ src/main/java/br/com/ifpe/conecta_vagas/api/vagas/VagasController.java
@@ -1,5 +1,9 @@
 package br.com.ifpe.conecta_vagas.api.vagas;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -8,8 +12,10 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import br.com.ifpe.conecta_vagas.modelo.recrutador.RecrutadorService;
 import br.com.ifpe.conecta_vagas.modelo.vagas.Vagas;
 import br.com.ifpe.conecta_vagas.modelo.vagas.VagasService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,6 +30,14 @@ import org.springframework.web.bind.annotation.PutMapping;
 public class VagasController {
     @Autowired
     private VagasService vagasService;
+    @Autowired
+    private RecrutadorService recrutadorService;
+
+    @GetMapping
+    public ResponseEntity<List<Vagas>> findAll(){
+        List<Vagas> todasVagas = vagasService.findAllVagas();
+        return new ResponseEntity<List<Vagas>>(todasVagas, HttpStatus.OK);
+    }
 
     @GetMapping("/{id}") // Uma única vaga
     public ResponseEntity<Vagas> findOne(@PathVariable("id") Long id) {
@@ -31,20 +45,36 @@ public class VagasController {
         return new ResponseEntity<Vagas>(vaga, HttpStatus.OK);
     }
 
-    @PostMapping("/{id}")
-    public ResponseEntity<Vagas> save(@PathVariable("id") Long id, @RequestBody @Valid VagasRequest request) {
-        Vagas vaga = this.vagasService.save(id, request.build());
+    @PostMapping
+    public ResponseEntity<Vagas> save(@RequestBody @Valid VagasRequest vagasRequest, HttpServletRequest request) {
+        Vagas vaga = this.vagasService.save(recrutadorService.obterRecrutadorLogado(request).getId(),
+                vagasRequest.build());
         return new ResponseEntity<Vagas>(vaga, HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Vagas> update(@PathVariable("id") Long id, @RequestBody @Valid VagasRequest request) {
-        Vagas vaga = this.vagasService.update(id, request.build());
+    public ResponseEntity<?> update(@PathVariable("id") Long id, @RequestBody @Valid VagasRequest vagasRequest,
+            HttpServletRequest request) {
+        boolean recrutadorPossuiVaga = recrutadorService.obterRecrutadorLogado(request).getVagas().stream()
+                .anyMatch(vaga -> vaga.getId().equals(id));
+        if (!recrutadorPossuiVaga) {
+            Map<String, Object> error = new HashMap<>();
+            error.put("message", "Você não tem permissão para alterar essa vaga");
+            return new ResponseEntity<Map<String, Object>>(error, HttpStatus.UNAUTHORIZED);
+        }
+        Vagas vaga = this.vagasService.update(id, vagasRequest.build());
         return new ResponseEntity<Vagas>(vaga, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> remove(@PathVariable("id") Long id) {
+    public ResponseEntity<?> remove(@PathVariable("id") Long id, HttpServletRequest request) {
+        boolean recrutadorPossuiVaga = recrutadorService.obterRecrutadorLogado(request).getVagas().stream()
+                .anyMatch(vaga -> vaga.getId().equals(id));
+        if (!recrutadorPossuiVaga) {
+            Map<String, Object> error = new HashMap<>();
+            error.put("message", "Você não tem permissão para alterar essa vaga");
+            return new ResponseEntity<Map<String, Object>>(error, HttpStatus.UNAUTHORIZED);
+        }
         this.vagasService.remove(id);
         return ResponseEntity.ok().build();
     }
